# Improve performance of SGEMM and STRMM on Arm Cortex-A53