… kernel/arm64/strmm_kernel_8x8_cortexa53.S