@craft-zhang 2020-05-18
> Thank you for the patch.
> 
> Does your optimized `sgemm_kernel_8x8` have better performance ( compared to old `sgemm_kernel_8x8` implementation )for other ARMv8 CPU (e.g. cortex-a73).

This optimization is specially customized according to cortexa53's pipe,  so i have changed it to only work for cortexa53.