@wjc404 2020-05-18
Interesting design. Once I noticed that the instruction "fmov vA.d[1], xB" (0<= A, B <32) does not compete with the execution of 128-bit fmla on cortex-A55, but significantly slows down that on cortex A76. Does OpenBLAS have a mechanism of dispatching different kernels to different cores in parallel calculation?