@craft-zhang 2020-05-18
> Interesting design. Once I noticed that the instruction "fmov vA.d[1], xB" (0<= A, B <32) does not interfere with the execution of 128-bit fmla on cortex-A55, but significantly slows down that on cortex A76. Does OpenBLAS have a mechanism of dispatching different kernels to different cores in parallel calculation?

If there is a way to calculate the threaded level3 blas that Big cores and little cores in parallelling, you may need a static load balancing strategy, which will make the code design very complicated