@martin-frbg 2020-05-18
There is currenly no mechanism in OpenBLAS to select different kinds of kernels onto different cores in a mixed configuration at runtime, and no active handling of big.little configurations in general. (Basically whichever core type gets scheduled first by the operating system "wins" in the autodetection). Adding this would indeed be quite complicated, and does not look particularly attractive given that only the ARMV8 architecture would currently make use of it.