@wjc404 2020-05-18
Thanks for your answers. 
I suspect that there's still room for improvement by reordering of some instructions:

.macro KERNEL8x8_I
  ld1	{v0.4s, v1.4s}, [pA], #32
  ld1	{v4.4s, v5.4s}, [pB], #32
  ldr   d2,  [pA], #8
  ldr   d6,  [pB], #8
  ldr   d3,  [pA, #8]
  ldr   d7,  [pB, #8]
  ldr   x20, [pA], #16
  fmul  v16.4s, v0.4s, v4.s[0]
  ldr   x24, [pB], #16
  fmul  v17.4s, v1.4s, v4.s[0]
  ldr   x21, [pA], #8
  fmul  v18.4s, v0.4s, v4.s[1]
  ldr   x25, [pB], #8
  fmul  v19.4s, v1.4s, v4.s[1]
  fmul  v20.4s, v0.4s, v4.s[2]
  fmul  v21.4s, v1.4s, v4.s[2]
  fmul  v22.4s, v0.4s, v4.s[3]
  fmul  v23.4s, v1.4s, v4.s[3]
  fmul  v24.4s, v0.4s, v5.s[0]
  fmul  v25.4s, v1.4s, v5.s[0]
  fmov  v2.d[1], x20
  fmul  v26.4s, v0.4s, v5.s[1]
  fmov  v3.d[1], x21
  fmul  v27.4s, v1.4s, v5.s[1]
  fmov  v6.d[1], x24
  fmul  v28.4s, v0.4s, v5.s[2]
  fmov  v7.d[1], x25
  fmul  v29.4s, v1.4s, v5.s[2]
  fmul  v30.4s, v0.4s, v5.s[3]
  fmul  v31.4s, v1.4s, v5.s[3]
.endm

.macro KERNEL8x8_M1
  ldr   d2, [pA], #32
  fmla  v16.4s, v0.4s, v4.s[0]
  ldr   d6,  [pB], #32
  fmla  v17.4s, v1.4s, v4.s[0]
  ldr   x20, [pA, #-24]
  fmla  v18.4s, v0.4s, v4.s[1]
  ldr   d3,  [pA, #-16]
  fmla  v19.4s, v1.4s, v4.s[1]
  ldr   x21, [pA, #-8]
  fmla  v20.4s, v0.4s, v4.s[2]
  ldr   x24, [pB, #-24]
  fmla  v21.4s, v1.4s, v4.s[2]
  ldr   d7,  [pB, #-16]
  fmla  v22.4s, v0.4s, v4.s[3]
  ldr   x25, [pB, #-8]
  fmla  v23.4s, v1.4s, v4.s[3]
  fmla  v24.4s, v0.4s, v5.s[0]
  fmov  v2.d[1], x20
  fmla  v25.4s, v1.4s, v5.s[0]
  fmla  v26.4s, v0.4s, v5.s[1]
  fmov  v3.d[1], x21
  fmla  v27.4s, v1.4s, v5.s[1]
  fmov  v6.d[1], x24
  fmla  v28.4s, v0.4s, v5.s[2]
  fmov  v7.d[1], x25
  fmla  v29.4s, v1.4s, v5.s[2]
  fmla  v30.4s, v0.4s, v5.s[3]
  fmla  v31.4s, v1.4s, v5.s[3]
.endm

.macro KERNEL8x8_M2
  ldr   d0,  [pA], #32
  fmla  v16.4s, v2.4s, v6.s[0]
  ldr   d4,  [pB], #32
  fmla  v17.4s, v3.4s, v6.s[0]
  ldr   x18, [pA, #-24]
  fmla  v18.4s, v2.4s, v6.s[1]
  ldr   d1,  [pA, #-16]
  fmla  v19.4s, v3.4s, v6.s[1]
  ldr   x19, [pA, #-8]
  fmla  v20.4s, v2.4s, v6.s[2]
  ldr   x22, [pB, #-24]
  fmla  v21.4s, v3.4s, v6.s[2]
  ldr   d5,  [pB, #-16]
  fmla  v22.4s, v2.4s, v6.s[3]
  ldr   x23, [pB, #-8]
  fmla  v23.4s, v3.4s, v6.s[3]
  fmla  v24.4s, v2.4s, v7.s[0]
  fmov  v0.d[1], x18
  fmla  v25.4s, v3.4s, v7.s[0]
  fmla  v26.4s, v2.4s, v7.s[1]
  fmov  v1.d[1], x19
  fmla  v27.4s, v3.4s, v7.s[1]
  fmov  v4.d[1], x22
  fmla  v28.4s, v2.4s, v7.s[2]
  fmov  v5.d[1], x23
  fmla  v29.4s, v3.4s, v7.s[2]
  fmla  v30.4s, v2.4s, v7.s[3]
  fmla  v31.4s, v3.4s, v7.s[3]
.endm

.macro KERNEL8x8_E
  fmla  v16.4s, v2.4s, v6.s[0]
  fmla  v17.4s, v3.4s, v6.s[0]
  fmla  v18.4s, v2.4s, v6.s[1]
  fmla  v19.4s, v3.4s, v6.s[1]
  fmla  v20.4s, v2.4s, v6.s[2]
  fmla  v21.4s, v3.4s, v6.s[2]
  fmla  v22.4s, v2.4s, v6.s[3]
  fmla  v23.4s, v3.4s, v6.s[3]
  fmla  v24.4s, v2.4s, v7.s[0]
  fmla  v25.4s, v3.4s, v7.s[0]
  fmla  v26.4s, v2.4s, v7.s[1]
  fmla  v27.4s, v3.4s, v7.s[1]
  fmla  v28.4s, v2.4s, v7.s[2]
  fmla  v29.4s, v3.4s, v7.s[2]
  fmla  v30.4s, v2.4s, v7.s[3]
  fmla  v31.4s, v3.4s, v7.s[3]
.endm