@craft-zhang 2020-05-18
Thanks for your advice and this is my mkernel benchmark result
before reordering:
Run on (6 X 1416 MHz CPU s)
sgemm_kernel_8x8/kernel_8x8/mr:8/nr:8/k:128/ldc:8        2310 ns         2310 ns       303036 items_per_second=7.09287G/s
sgemm_kernel_8x8/kernel_8x8/mr:8/nr:8/k:256/ldc:8        4496 ns         4496 ns       155735 items_per_second=7.28775G/s
sgemm_kernel_8x8/kernel_8x8/mr:8/nr:8/k:512/ldc:8        9282 ns         9282 ns        75405 items_per_second=7.06079G/s
sgemm_kernel_8x8/kernel_8x8/mr:8/nr:8/k:1024/ldc:8      20103 ns        20102 ns        34971 items_per_second=6.52021G/s
sgemm_kernel_8x8/kernel_8x8/mr:8/nr:8/k:2048/ldc:8      39452 ns        39432 ns        17773 items_per_second=6.64797G/s
after reordering:
Run on (6 X 1416 MHz CPU s)
sgemm_kernel_8x8/kernel_8x8/mr:8/nr:8/k:128/ldc:8        2749 ns         2749 ns       247725 items_per_second=5.96G/s
sgemm_kernel_8x8/kernel_8x8/mr:8/nr:8/k:256/ldc:8        5378 ns         5378 ns       130119 items_per_second=6.09242G/s
sgemm_kernel_8x8/kernel_8x8/mr:8/nr:8/k:512/ldc:8       11226 ns        11226 ns        62352 items_per_second=5.838G/s
sgemm_kernel_8x8/kernel_8x8/mr:8/nr:8/k:1024/ldc:8      23292 ns        23292 ns        30037 items_per_second=5.62729G/s
sgemm_kernel_8x8/kernel_8x8/mr:8/nr:8/k:2048/ldc:8      45528 ns        45527 ns        15373 items_per_second=5.75803G/s

Since there is a difference between a53 and a55, fmov cannot dual-i
ssue with fmla on a53, reordering may reduce performance.