@wjc404 2020-05-18
How about storing elements from matrix B in 4 64-bit neon registers per iteration ? I think this can eliminate half of fmov instructions.