@craft-zhang 2020-05-18
Sorry about my English, hope i can Explain clearly.
Firstly, a53's fmla pipeline is after 4 cycles issues add, In other words, fmla cannot be dual-issued with any other simd instruction. Once a simd instruction is inserted, it will wait for 4cycle to be issued. In order to amortize, we need the mkernel to be the maximum calculation access memory ratio, that is 8x8.
Secondly, fmov can be dual-issued with ld1 and fmla can be dual-issued with ldr gr. So laterly we only need to ensure that ld1 and fmov are fully dual-issued.