@wjc404 2020-05-18
Sorry for my stupid suggestions. So on A53 there're no instructions operating on neon registers (read or write) dual-issuable with 128-bit fmla, is that correct?