@craft-zhang 2020-05-18
This is something worth discussing. According to some of my experiments and inferences, it should be like this. But i can't guarantee. : )