@wjc404 2020-05-21
Just one more stupid question, how many cycles are required for a post-indexed load instruction to update address register on A53 (from issue to being available for subsequent load) ?